Convert game-launcher and related stuff to TS
