+++ src/game/game.ts
@@ -5,6 +5,7 @@ export type GamePlayerOptionsInput = {
   players: Omit<PlayerOptionsInput, 'autosound' | 'tutorial'>[]
   audioInputLatency: number
   tutorial: boolean
+  soundVolume: number
 }
 
 /**
