+++ src/logger/global.d.ts
@@ -1,0 +1,1 @@
+var BemuseLogger: typeof import('./index')
